+++ pkg/scalers/mssql_scaler_test.go
@@ -16,6 +16,7 @@ type parseMSSQLMetadataTestData struct {
 	authParams               map[string]string
 	expectedError            string
 	expectedConnectionString string
+	expectedDriverName       string
 	expectedMetricName       string
 }
 
@@ -88,6 +89,30 @@ var testMSSQLMetadata = []parseMSSQLMetadataTestData{
 		expectedConnectionString: "sqlserver://user3:Password%233@example.database.windows.net",
 	},
 	{
+		name:                     "no driver selected; default to sqlserver",
+		metadata:                 map[string]string{"query": "SELECT 1", "targetValue": "1", "host": "example.database.windows.net", "username": "user4"},
+		resolvedEnv:              map[string]string{},
+		authParams:               map[string]string{"password": "Password#4"},
+		expectedDriverName:       "sqlserver",
+		expectedConnectionString: "sqlserver://user4:Password%234@example.database.windows.net",
+	},
+	{
+		name:                     "sqlserver driver selected",
+		metadata:                 map[string]string{"query": "SELECT 1", "targetValue": "1", "host": "example.database.windows.net", "username": "user5"},
+		resolvedEnv:              map[string]string{},
+		authParams:               map[string]string{"password": "Password#5", "driverName": "sqlserver"},
+		expectedDriverName:       "sqlserver",
+		expectedConnectionString: "sqlserver://user5:Password%235@example.database.windows.net",
+	},
+	{
+		name:                     "azuresql driver selected",
+		metadata:                 map[string]string{"query": "SELECT 1", "targetValue": "1", "host": "example.database.windows.net", "username": "user6"},
+		resolvedEnv:              map[string]string{},
+		authParams:               map[string]string{"password": "Password#6", "driverName": "azuresql"},
+		expectedDriverName:       "azuresql",
+		expectedConnectionString: "sqlserver://user6:Password%236@example.database.windows.net",
+	},
+	{
 		name:          "Error: missing query",
 		metadata:      map[string]string{"targetValue": "1"},
 		resolvedEnv:   map[string]string{},
@@ -108,6 +133,13 @@ var testMSSQLMetadata = []parseMSSQLMetadataTestData{
 		authParams:    map[string]string{},
 		expectedError: "must provide either connectionstring or host",
 	},
+	{
+		name:          "Error:  invalid driver",
+		metadata:      map[string]string{"query": "SELECT 1", "targetValue": "1", "host": "example.database.windows.net", "username": "user7"},
+		resolvedEnv:   map[string]string{},
+		authParams:    map[string]string{"password": "Password#7", "driverName": "fake"},
+		expectedError: "parameter \"driverName\" value \"fake\" must be one of [sqlserver azuresql]",
+	},
 }
 
 func TestParseMSSQLMetadata(t *testing.T) {
@@ -126,7 +158,37 @@ func TestParseMSSQLMetadata(t *testing.T) {
 			} else {
 				assert.NoError(t, err)
 				assert.NotNil(t, meta)
+
+				if testData.expectedDriverName != "" {
+					assert.Equal(t, testData.expectedDriverName, meta.DriverName)
+				}
 			}
+		})
+	}
+}
+
+func TestMSSQLConnectionStringGeneration(t *testing.T) {
+	for _, testData := range testMSSQLMetadata {
+		t.Run(testData.name, func(t *testing.T) {
+			if testData.expectedError != "" {
+				return
+			}
+
+			meta, err := parseMSSQLMetadata(&scalersconfig.ScalerConfig{
+				TriggerMetadata: testData.metadata,
+				ResolvedEnv:     testData.resolvedEnv,
+				AuthParams:      testData.authParams,
+			})
+
+			assert.NoError(t, err)
+
+			mockMSSQLScaler := mssqlScaler{
+				metadata: meta,
+			}
+
+			connStr := getMSSQLConnectionString(&mockMSSQLScaler)
+
+			assert.Equal(t, testData.expectedConnectionString, connStr)
 		})
 	}
 }